refactor: organize account route tests by workflow
