+++ backend/tests/routes/accounts/__init__.py
@@ -1,0 +1,1 @@
+"""Account route tests"""
