+++ backend/tests/routes/accounts/test_account_deletion.py
@@ -1,0 +1,69 @@
+
+
+from tests.routes.accounts._account_helpers import (
+    NONEXISTENT_ID,
+    _create_second_user,
+)
+from tests.routes.support import _create_account, _create_user, _get_auth_header
+
+# --- DELETE /accounts/{account_id} ---
+
+
+async def test_delete_account_returns_204(client):
+    """DELETE removes the account and returns 204."""
+    signup_resp = await _create_user(client)
+    headers = _get_auth_header(signup_resp)
+    create_resp = await _create_account(client, headers)
+    account_id = create_resp.json()["id"]
+
+    resp = await client.delete(f"/accounts/{account_id}", headers=headers)
+
+    assert resp.status_code == 204
+
+    # Verify account is gone
+    get_resp = await client.get(f"/accounts/{account_id}", headers=headers)
+    assert get_resp.status_code == 404
+
+
+async def test_delete_account_not_found_returns_404(client):
+    """DELETE non-existent account returns 404."""
+    signup_resp = await _create_user(client)
+    headers = _get_auth_header(signup_resp)
+
+    resp = await client.delete(f"/accounts/{NONEXISTENT_ID}", headers=headers)
+
+    assert resp.status_code == 404
+
+
+async def test_delete_account_other_user_returns_404(client):
+    """Deleting another user's account returns 404."""
+    signup_resp = await _create_user(client)
+    headers = _get_auth_header(signup_resp)
+    create_resp = await _create_account(client, headers)
+    account_id = create_resp.json()["id"]
+
+    other_headers = _get_auth_header(await _create_second_user(client))
+
+    resp = await client.delete(f"/accounts/{account_id}", headers=other_headers)
+
+    assert resp.status_code == 404
+
+
+async def test_delete_account_without_auth_returns_401(client):
+    """DELETE /accounts/{id} without an Authorization header returns 401."""
+    resp = await client.delete(f"/accounts/{NONEXISTENT_ID}")
+    assert resp.status_code == 401
+
+
+async def test_double_delete_returns_404_on_second(client):
+    """Deleting the same account twice returns 204 then 404."""
+    signup_resp = await _create_user(client)
+    headers = _get_auth_header(signup_resp)
+    create_resp = await _create_account(client, headers)
+    account_id = create_resp.json()["id"]
+
+    resp1 = await client.delete(f"/accounts/{account_id}", headers=headers)
+    resp2 = await client.delete(f"/accounts/{account_id}", headers=headers)
+
+    assert resp1.status_code == 204
+    assert resp2.status_code == 404
