+++ backend/tests/routes/accounts/test_balance_snapshot_events.py
@@ -1,60 +1,19 @@
 """Route tests for the account balance snapshot endpoints and lifecycle hooks."""
 import uuid
-from datetime import date, datetime
-from zoneinfo import ZoneInfo
+from datetime import date
 
 from sqlalchemy import select
 
 from app.models.account import AccountBalanceSnapshot
-from app.models.currency import Currency
 from tests.conftest import TestSession
-from tests.routes.conftest import _create_account, _create_user, _get_auth_header
-
-# --- Helpers ---
-
-
-
-def _creation_day(account_resp):
-    """Return the account owner's local creation date from the API timestamp."""
-    return datetime.fromisoformat(account_resp.json()["created_at"]).astimezone(ZoneInfo("America/Toronto")).date()
-
-
-async def _get_snapshots_for(account_id):
-    """Query the DB directly for an account's balance snapshots ordered by ts."""
-    async with TestSession() as session:
-        result = await session.execute(
-            select(AccountBalanceSnapshot)
-            .where(AccountBalanceSnapshot.account_id == account_id)
-            .order_by(AccountBalanceSnapshot.dt),
-        )
-        return list(result.scalars().all())
-
-
-async def _seed_usd_currency():
-    """Insert the USD currency row needed for multi-currency transaction tests."""
-    async with TestSession() as session:
-        session.add(Currency(id="USD", name="US Dollar", symbol="$", minor_unit_exponent=2))
-        await session.commit()
-
-
-async def _create_category(client, headers, **overrides):
-    """Create an expense category via POST /categories."""
-    payload = {"name": "Test Groceries", "kind": "expense", **overrides}
-    return await client.post("/categories", json=payload, headers=headers)
-
-
-async def _create_transaction(client, headers, account_id, category_id, **overrides):
-    """Create a transaction via POST /transactions."""
-    payload = {
-        "account_id": account_id,
-        "category_id": category_id,
-        "dt": "2026-03-15",
-        "amount": -5000,
-        "currency": "CAD",
-        **overrides,
-    }
-    return await client.post("/transactions", json=payload, headers=headers)
-
+from tests.routes.accounts._balance_snapshot_helpers import (
+    _create_category,
+    _create_transaction,
+    _creation_day,
+    _get_snapshots_for,
+    _seed_usd_currency,
+)
+from tests.routes.support import _create_account, _create_user, _get_auth_header
 
 # --- Zero snapshot seeding on account creation ---
 
@@ -914,447 +873,3 @@ async def test_delete_group_account_transaction_recomputes_snapshots(client):
     assert date(2026, 3, 15) not in snapshot_map
     assert snapshot_map[creation_day] == 0
     assert len(snapshots) == 1
-
-
-# --- Helpers for the snapshots endpoint ---
-
-NONEXISTENT_ID = "00000000-0000-0000-0000-000000000000"
-
-
-async def _create_second_user(client):
-    """Sign up a second user and return (auth_headers, user_id)."""
-    resp = await client.post("/auth/signup", json={
-        "email": "other@example.com",
-        "password": "securepassword123",
-        "first_name": "Other",
-        "tz": "America/Toronto",
-        "base_currency": "CAD",
-    })
-    return _get_auth_header(resp), resp.json()["user"]["id"]
-
-
-async def _create_group(client, headers):
-    """Create a group and return its id."""
-    resp = await client.post("/groups", json={"name": "Smith Family"}, headers=headers)
-    return resp.json()["id"]
-
-
-async def _grant_account_permission(client, admin_headers, account_id, user_id, level):
-    """Grant a user a permission level on a group account."""
-    return await client.post(
-        f"/accounts/{account_id}/permissions",
-        json={"user_id": user_id, "level": level},
-        headers=admin_headers,
-    )
-
-
-async def _seed_three_day_history(client, headers, account_id):
-    """Create transactions on 3/1, 3/5, and 3/10 so the account has 3 snapshots."""
-    cat_resp = await _create_category(client, headers)
-    category_id = cat_resp.json()["id"]
-    await _create_transaction(
-        client, headers, account_id, category_id,
-        dt="2026-03-01", amount=1000,
-    )
-    await _create_transaction(
-        client, headers, account_id, category_id,
-        dt="2026-03-05", amount=2000,
-    )
-    await _create_transaction(
-        client, headers, account_id, category_id,
-        dt="2026-03-10", amount=3000,
-    )
-
-
-# --- GET /accounts/{account_id}/snapshots — listing and date filters ---
-
-
-async def test_list_snapshots_returns_all_in_ascending_order(client):
-    """The endpoint returns every snapshot for the account ordered dt ascending."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    await _seed_three_day_history(client, headers, account_id)
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=headers)
-    assert resp.status_code == 200
-
-    snapshots = resp.json()
-    assert len(snapshots) == 3
-    timestamps = [s["dt"] for s in snapshots]
-    assert timestamps == sorted(timestamps)
-    assert snapshots[0]["balance"] == 1000
-    assert snapshots[1]["balance"] == 3000
-    assert snapshots[2]["balance"] == 6000
-
-
-async def test_list_snapshots_filters_by_from_date(client):
-    """from_date excludes snapshots strictly before the bound (inclusive boundary).
-
-    Passing from_date equal to a snapshot's ts includes that snapshot.
-    """
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    await _seed_three_day_history(client, headers, account_id)
-
-    resp = await client.get(
-        f"/accounts/{account_id}/snapshots",
-        params={"from_date": "2026-03-05"},
-        headers=headers,
-    )
-    assert resp.status_code == 200
-
-    snapshots = resp.json()
-    assert len(snapshots) == 2
-    assert snapshots[0]["balance"] == 3000
-    assert snapshots[1]["balance"] == 6000
-
-
-async def test_list_snapshots_filters_by_to_date(client):
-    """to_date excludes snapshots strictly after the bound (inclusive boundary).
-
-    Passing to_date equal to a snapshot's ts includes that snapshot.
-    """
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    await _seed_three_day_history(client, headers, account_id)
-
-    resp = await client.get(
-        f"/accounts/{account_id}/snapshots",
-        params={"to_date": "2026-03-05"},
-        headers=headers,
-    )
-    assert resp.status_code == 200
-
-    snapshots = resp.json()
-    assert len(snapshots) == 2
-    assert snapshots[0]["balance"] == 1000
-    assert snapshots[1]["balance"] == 3000
-
-
-async def test_list_snapshots_filters_by_both_date_bounds(client):
-    """Both bounds combined return only snapshots inside the inclusive range."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    await _seed_three_day_history(client, headers, account_id)
-
-    resp = await client.get(
-        f"/accounts/{account_id}/snapshots",
-        params={
-            "from_date": "2026-03-04",
-            "to_date": "2026-03-06",
-        },
-        headers=headers,
-    )
-    assert resp.status_code == 200
-
-    snapshots = resp.json()
-    assert len(snapshots) == 1
-    assert snapshots[0]["balance"] == 3000
-
-
-async def test_list_snapshots_with_zero_width_date_range_returns_only_that_day(client):
-    """from_date == to_date covering one snapshot's day returns exactly that snapshot."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    await _seed_three_day_history(client, headers, account_id)
-
-    resp = await client.get(
-        f"/accounts/{account_id}/snapshots",
-        params={
-            "from_date": "2026-03-05",
-            "to_date": "2026-03-05",
-        },
-        headers=headers,
-    )
-    assert resp.status_code == 200
-
-    snapshots = resp.json()
-    assert len(snapshots) == 1
-    assert snapshots[0]["balance"] == 3000
-
-
-async def test_list_snapshots_returns_empty_when_no_snapshots_in_range(client):
-    """A date range that excludes all snapshots returns an empty list, not 404."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    await _seed_three_day_history(client, headers, account_id)
-
-    resp = await client.get(
-        f"/accounts/{account_id}/snapshots",
-        params={
-            "from_date": "2027-01-01",
-            "to_date": "2027-12-31",
-        },
-        headers=headers,
-    )
-    assert resp.status_code == 200
-    assert resp.json() == []
-
-
-async def test_list_snapshots_returns_zero_anchor_for_new_account(client):
-    """A brand-new account with no transactions returns just the zero anchor on its creation day."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    expected_anchor_dt = _creation_day(account_resp)
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=headers)
-    assert resp.status_code == 200
-
-    snapshots = resp.json()
-    assert len(snapshots) == 1
-    assert snapshots[0]["balance"] == 0
-    assert date.fromisoformat(snapshots[0]["dt"]) == expected_anchor_dt
-
-
-async def test_list_snapshots_on_closed_account_still_returns_history(client):
-    """Read-only endpoint must return snapshots even after the account is closed.
-
-    Closed accounts are still meaningful for historical balance charts; the
-    handler intentionally does NOT pass require_open=True to check_account_access.
-    """
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-    await _seed_three_day_history(client, headers, account_id)
-
-    # Close the account
-    close_resp = await client.patch(
-        f"/accounts/{account_id}",
-        json={"closed_at": "2026-04-01"},
-        headers=headers,
-    )
-    assert close_resp.status_code == 200
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=headers)
-    assert resp.status_code == 200
-    assert len(resp.json()) == 3
-
-
-# --- GET /accounts/{account_id}/snapshots — validation ---
-
-
-async def test_list_snapshots_with_inverted_date_range_returns_422(client):
-    """from_date later than to_date is rejected."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-
-    resp = await client.get(
-        f"/accounts/{account_id}/snapshots",
-        params={
-            "from_date": "2026-04-01",
-            "to_date": "2026-03-01",
-        },
-        headers=headers,
-    )
-    assert resp.status_code == 422
-
-
-async def test_list_snapshots_with_invalid_date_format_returns_422(client):
-    """A malformed date string is rejected by FastAPI's query validation."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-
-    resp = await client.get(
-        f"/accounts/{account_id}/snapshots",
-        params={"from_date": "not-a-date"},
-        headers=headers,
-    )
-    assert resp.status_code == 422
-
-
-# --- GET /accounts/{account_id}/snapshots — auth and permissions ---
-
-
-async def test_list_snapshots_unauthenticated_returns_401(client):
-    """Anonymous requests are rejected before reaching the handler."""
-    resp = await client.get(f"/accounts/{NONEXISTENT_ID}/snapshots")
-    assert resp.status_code == 401
-
-
-async def test_list_snapshots_unknown_account_returns_404(client):
-    """A nonexistent account UUID returns 404."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    resp = await client.get(f"/accounts/{NONEXISTENT_ID}/snapshots", headers=headers)
-    assert resp.status_code == 404
-
-
-async def test_list_snapshots_other_users_personal_account_returns_404(client):
-    """A second user cannot enumerate the existence of a personal account they don't own."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-
-    other_headers, _ = await _create_second_user(client)
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=other_headers)
-    assert resp.status_code == 404
-
-
-async def test_list_snapshots_personal_owner_can_read_own_account(client):
-    """A personal account owner can read their own snapshots."""
-    signup_resp = await _create_user(client)
-    headers = _get_auth_header(signup_resp)
-
-    account_resp = await _create_account(client, headers)
-    account_id = account_resp.json()["id"]
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=headers)
-    assert resp.status_code == 200
-    assert len(resp.json()) == 1
-    assert resp.json()[0]["balance"] == 0
-
-
-async def test_list_snapshots_group_admin_can_read_group_account(client):
-    """A group admin has implicit access to read group account snapshots."""
-    signup_resp = await _create_user(client)
-    admin_headers = _get_auth_header(signup_resp)
-
-    group_id = await _create_group(client, admin_headers)
-    account_resp = await _create_account(client, admin_headers, group_id=group_id)
-    account_id = account_resp.json()["id"]
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=admin_headers)
-    assert resp.status_code == 200
-    assert len(resp.json()) == 1
-    assert resp.json()[0]["balance"] == 0
-
-
-async def test_list_snapshots_non_group_user_returns_404(client):
-    """A user who is not a member of the account's group at all gets 404.
-
-    Distinct code path from "group member without permission": this user
-    fails the membership lookup before any AccountPermission check.
-    """
-    signup_resp = await _create_user(client)
-    admin_headers = _get_auth_header(signup_resp)
-
-    group_id = await _create_group(client, admin_headers)
-    account_resp = await _create_account(client, admin_headers, group_id=group_id)
-    account_id = account_resp.json()["id"]
-
-    # Other user is intentionally NOT added to the group
-    other_headers, _ = await _create_second_user(client)
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=other_headers)
-    assert resp.status_code == 404
-
-
-async def test_list_snapshots_group_member_with_read_permission_can_access(client):
-    """A group member granted explicit READ on the account can read its snapshots."""
-    signup_resp = await _create_user(client)
-    admin_headers = _get_auth_header(signup_resp)
-
-    group_id = await _create_group(client, admin_headers)
-    account_resp = await _create_account(client, admin_headers, group_id=group_id)
-    account_id = account_resp.json()["id"]
-
-    member_headers, member_user_id = await _create_second_user(client)
-    await client.post(
-        f"/groups/{group_id}/members",
-        json={"user_id": member_user_id},
-        headers=admin_headers,
-    )
-    await _grant_account_permission(client, admin_headers, account_id, member_user_id, "read")
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=member_headers)
-    assert resp.status_code == 200
-    assert len(resp.json()) == 1
-    assert resp.json()[0]["balance"] == 0
-
-
-async def test_list_snapshots_group_member_without_permission_returns_404(client):
-    """A group member with no explicit permission on the account gets 404, not 403."""
-    signup_resp = await _create_user(client)
-    admin_headers = _get_auth_header(signup_resp)
-
-    group_id = await _create_group(client, admin_headers)
-    account_resp = await _create_account(client, admin_headers, group_id=group_id)
-    account_id = account_resp.json()["id"]
-
-    member_headers, member_user_id = await _create_second_user(client)
-    await client.post(
-        f"/groups/{group_id}/members",
-        json={"user_id": member_user_id},
-        headers=admin_headers,
-    )
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=member_headers)
-    assert resp.status_code == 404
-
-
-async def test_list_snapshots_group_member_with_write_permission_can_read(client):
-    """WRITE access implies READ — a member with WRITE can read snapshots."""
-    signup_resp = await _create_user(client)
-    admin_headers = _get_auth_header(signup_resp)
-
-    group_id = await _create_group(client, admin_headers)
-    account_resp = await _create_account(client, admin_headers, group_id=group_id)
-    account_id = account_resp.json()["id"]
-
-    member_headers, member_user_id = await _create_second_user(client)
-    await client.post(
-        f"/groups/{group_id}/members",
-        json={"user_id": member_user_id},
-        headers=admin_headers,
-    )
-    await _grant_account_permission(client, admin_headers, account_id, member_user_id, "write")
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=member_headers)
-    assert resp.status_code == 200
-    assert len(resp.json()) == 1
-    assert resp.json()[0]["balance"] == 0
-
-
-async def test_list_snapshots_group_member_with_admin_permission_can_read(client):
-    """ADMIN access also implies READ — locks in the WRITE < ADMIN ladder ordering."""
-    signup_resp = await _create_user(client)
-    admin_headers = _get_auth_header(signup_resp)
-
-    group_id = await _create_group(client, admin_headers)
-    account_resp = await _create_account(client, admin_headers, group_id=group_id)
-    account_id = account_resp.json()["id"]
-
-    member_headers, member_user_id = await _create_second_user(client)
-    await client.post(
-        f"/groups/{group_id}/members",
-        json={"user_id": member_user_id},
-        headers=admin_headers,
-    )
-    await _grant_account_permission(client, admin_headers, account_id, member_user_id, "admin")
-
-    resp = await client.get(f"/accounts/{account_id}/snapshots", headers=member_headers)
-    assert resp.status_code == 200
-    assert len(resp.json()) == 1
-    assert resp.json()[0]["balance"] == 0
