+++ backend/tests/routes/accounts/test_spending_breakdown.py
@@ -3,7 +3,7 @@ import importlib
 from datetime import UTC, date, datetime, timedelta
 from zoneinfo import ZoneInfo
 
-from tests.routes.conftest import _create_account, _create_user, _get_auth_header
+from tests.routes.support import _create_account, _create_user, _get_auth_header
 
 # --- Helpers ---
