+++ backend/tests/routes/accounts/test_tax_advantaged_category_links.py
@@ -1,6 +1,6 @@
 from app.models.currency import Currency
 from tests.conftest import TestSession
-from tests.routes.conftest import _create_account, _create_user, _get_auth_header
+from tests.routes.support import _create_account, _create_user, _get_auth_header
 
 NONEXISTENT_ID = "00000000-0000-0000-0000-000000000000"
